Rename declaration-semicolon-* and add renaming docs
