+++ docs/rules.md
@@ -1,5 +1,6 @@
 # Rules
 
+* [`at-rule-empty-line-before`](../src/rules/at-rule-empty-line-before/README.md): Require or disallow an empty line before @rules.
 * [`at-rule-no-vendor-prefix`](../src/rules/at-rule-no-vendor-prefix/README.md): Disallow vendor prefixes for @rules.
 * [`block-closing-brace-newline-after`](../src/rules/block-closing-brace-newline-after/README.md): Require or disallow a newline after the closing brace of blocks.
 * [`block-closing-brace-newline-before`](../src/rules/block-closing-brace-newline-before/README.md): Require or disallow a newline before the closing brace blocks.
@@ -11,6 +12,7 @@
 * [`block-opening-brace-space-after`](../src/rules/block-opening-brace-space-after/README.md): Require or disallow a space after the opening brace of blocks.
 * [`block-opening-brace-space-before`](../src/rules/block-opening-brace-space-before/README.md): Require or disallow a space before the opening brace of blocks.
 * [`color-no-invalid-hex`](../src/rules/color-no-invalid-hex/README.md): Disallow invalid hex colors.
+* [`comment-empty-line-before`](../src/rules/comment-empty-line-before/README.md): Require or disallow an empty line before comments.
 * [`declaration-bang-space-after`](../src/rules/declaration-bang-space-after/README.md): Require or disallow a space after the bang of declarations.
 * [`declaration-bang-space-before`](../src/rules/declaration-bang-space-before/README.md): Require or disallow a space before the bang of declarations.
 * [`declaration-colon-space-after`](../src/rules/declaration-colon-space-after/README.md): Require or disallow a space after the colon of declarations.
@@ -33,14 +35,20 @@
 * [`media-query-parentheses-inside-space`](../src/rules/media-query-parentheses-inside-space/README.md): Require or disallow a single space on the inside of the parentheses within media queries.
 * [`no-eol-whitespace`](../src/rules/no-eol-whitespace/README.md): Disallow end-of-line whitespace.
 * [`no-missing-eof-newline`](../src/rules/no-missing-eof-newline/README.md): Disallow missing end-of-file newline.
+* [`no-multiple-empty-lines`](../src/rules/no-multiple-empty-lines/README.md): Disallow multiple empty lines.
 * [`number-leading-zero`](../src/rules/number-leading-zero/README.md): Require or disallow a leading zero for fractional numbers less than 1.
 * [`number-no-trailing-zeros`](../src/rules/number-no-trailing-zeros/README.md): Disallow trailing zeros within numbers.
 * [`property-no-vendor-prefix`](../src/rules/property-no-vendor-prefix/README.md): Disallow vendor prefixes for properties.
 * [`property-value-no-vendor-prefix`](../src/rules/property-value-no-vendor-prefix/README.md): Disallow vendor prefixes for property values.
 * [`root-no-standard-properties`](../src/rules/root-no-standard-properties/README.md): Disallow standard properties inside `:root` selectors.
+* [`rule-nested-empty-line-before`](../src/rules/rule-nested-empty-line-before/README.md): Require or disallow an empty line before nested rules.
 * [`rule-no-duplicate-properties`](../src/rules/rule-no-duplicate-properties/README.md): Disallow duplicate properties within rules.
 * [`rule-no-single-line`](../src/rules/rule-no-single-line/README.md): Disallow single-line rules.
 * [`rule-non-nested-empty-line-before`](../src/rules/rule-non-nested-empty-line-before/README.md): Require or disallow an empty line before non-nested rules.
+* [`rule-semicolon-newline-after`](../src/rules/rule-semicolon-newline-after/README.md): Require or disallow a newline after the semicolons of rules.
+* [`rule-semicolon-newline-before`](../src/rules/rule-semicolon-newline-before/README.md): Require or disallow a newline before the semicolons of rules.
+* [`rule-semicolon-space-after`](../src/rules/rule-semicolon-space-after/README.md): Require or disallow a space after the semicolons of rules.
+* [`rule-semicolon-space-before`](../src/rules/rule-semicolon-space-before/README.md): Require or disallow a space before the semicolons of rules.
 * [`rule-trailing-semicolon`](../src/rules/rule-trailing-semicolon/README.md): Require or disallow a trailing semicolon within rules.
 * [`selector-combinator-space-after`](../src/rules/selector-combinator-space-after/README.md): Require or disallow a space after the combinator of selectors.
 * [`selector-combinator-space-before`](../src/rules/selector-combinator-space-before/README.md): Require or disallow a space before the combinator of selectors.
