+++ src/rules/rule-nested-empty-line-before/README.md
@@ -1,0 +1,120 @@
+# rule-nested-empty-line-before
+
+Require or disallow an empty line before nested rules.
+
+```css
+    @media {
+           /* ← */
+      a {} /* ↑ */
+    }      /* ↑ */
+/**           ↑  
+ *    This line */
+```
+
+## Options
+
+`string`: `"always"|"never"|"always-multi-line"|"never-multi-line"`
+
+### `"always"`
+
+There *must always* be an empty line before rules.
+
+The following patterns are considered warnings:
+
+```css
+@media { a {} }
+```
+
+```css
+@media { 
+  a {} 
+}
+```
+
+The following patterns are *not* considered warnings:
+
+```css
+@media {
+
+  a {} 
+}
+```
+
+### `"never"`
+
+There *must never* be an empty before rules.
+
+The following patterns are considered warnings:
+
+```css
+@media { a {} }
+```
+
+```css
+@media { 
+
+  a {} 
+}
+```
+
+The following patterns are *not* considered warnings:
+
+```css
+@media { 
+  a {} 
+}
+```
+
+### `"always-multi-line"`
+
+There *must always* be an empty line before multi-line rules.
+
+The following patterns are considered warnings:
+
+```css
+@media { 
+  a {
+    color: pink;
+    top: 0;
+  } 
+}
+```
+
+The following patterns are *not* considered warnings:
+
+```css
+@media {
+
+  a {
+    color: pink;
+    top: 0;
+  } 
+}
+```
+
+### `"never-multi-line"`
+
+There *must never* be an empty line before multi-line rules.
+
+The following patterns are considered warnings:
+
+```css
+@media {
+
+  a {
+    color: pink;
+    top: 0;
+  } 
+}
+```
+
+The following patterns are *not* considered warnings:
+
+```css
+@media {
+  a {
+    color: pink;
+    top: 0;
+  } 
+}
+```
