+++ src/rules/rule-semicolon-newline-after/__tests__/index.js
