+++ src/rules/rule-semicolon-newline-before/__tests__/index.js
