+++ src/rules/rule-semicolon-space-after/__tests__/index.js
