+++ src/rules/rule-semicolon-space-before/__tests__/index.js
